# frontend/openapi-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HamzaTex API",
    "description": "Textile ERP — staff mobile + admin backend. All endpoints require a JWT Bearer token except Auth.",
    "version": "v1"
  },
  "paths": {
...
    "/api/Invoice": {
      "post": {
        "tags": ["Invoice"],
        "summary": "Create a standalone invoice.",
        "operationId": "invoice_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCreateViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCreateViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCreateViewModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDtoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Invoice"],
        "summary": "Get all invoices paginated. Admin only.",
        "operationId": "invoice_getAll",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDtoPagedListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Invoice/{id}": {
      "get": {
        "tags": ["Invoice"],
        "summary": "Get a single invoice with lines and linked transactions.",
        "operationId": "invoice_getById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailDtoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Invoice"],
        "summary": "Update invoice status, dueDate, notes, totalAmount, and/or lines.",
        "operationId": "invoice_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceUpdateViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceUpdateViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceUpdateViewModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDtoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Invoice"],
        "summary": "Delete a Draft invoice created within the last year. Admin only.",
        "operationId": "invoice_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
    "/api/Invoice/by-client/{clientId}": {
      "get": {
        "tags": ["Invoice"],
        "summary": "Get all invoices for a client with aggregate stats.",
        "operationId": "invoice_getByClient",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceByClientDtoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
    "/api/Invoice/filtered": {
      "get": {
        "tags": ["Invoice"],
        "summary": "Filter invoices by statusId, clientId, dateFrom, dateTo.",
        "operationId": "invoice_getFiltered",
        "parameters": [
          {
            "name": "statusId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDtoListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Invoice/pdf": {
      "get": {
        "tags": ["Invoice"],
        "summary": "Export all invoices as a branded PDF with stat cards.",
        "operationId": "invoice_getAllPdf",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Invoice/{id}/pdf": {
      "get": {
        "tags": ["Invoice"],
        "summary": "Export a single invoice as a formal branded PDF.",
        "operationId": "invoice_getByIdPdf",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
    "/api/Invoice/by-client/{clientId}/pdf": {
      "get": {
        "tags": ["Invoice"],
        "summary": "Export all invoices for a client as a PDF with aggregate stat cards.",
        "operationId": "invoice_getByClientPdf",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
    "/api/Invoice/filtered/pdf": {
      "get": {
        "tags": ["Invoice"],
        "summary": "Export filtered invoices as a PDF.",
        "operationId": "invoice_getFilteredPdf",
        "parameters": [
          {
            "name": "statusId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InvoiceByClientDtoResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "example": "Operation completed successfully"
          },
          "data": {
            "$ref": "#/components/schemas/InvoiceByClientDto"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InvoiceCreateViewModel": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "example": 1
          },
          "purchaseId": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "example": 1
          },
          "clientId": {
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "string"
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "example": "Optional notes here"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineCreateViewModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "InvoiceDetailDtoResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "example": "Operation completed successfully"
          },
          "data": {
            "$ref": "#/components/schemas/InvoiceDetailDto"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "InvoiceDtoListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "example": "Operation completed successfully"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceDto"
            },
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "InvoiceDtoPagedListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "example": "Operation completed successfully"
          },
          "data": {
            "$ref": "#/components/schemas/InvoiceDtoPagedList"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InvoiceDtoResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "example": "Operation completed successfully"
          },
          "data": {
            "$ref": "#/components/schemas/InvoiceDto"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InvoiceLineCreateViewModel": {
        "type": "object",
        "properties": {
          "productName": {
            "type": "string",
            "nullable": true,
            "example": "string"
          },
          "qty": {
            "type": "number",
            "format": "double",
            "example": 50
          },
          "unitPrice": {
            "type": "number",
            "format": "double",
            "example": 150000
          }
        },
        "additionalProperties": false
      },
...
      "InvoiceUpdateViewModel": {
        "type": "object",
        "properties": {
          "invoiceStatusId": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "example": 1
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "example": "Optional notes here"
          },
          "totalAmount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineCreateViewModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Response": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "example": "Operation completed successfully"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter your JWT access token (without the 'Bearer ' prefix).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}